Consolidate summaries
