+++ src/app/store/promotion-type.ts
@@ -1,0 +1,4 @@
+export interface PromotionType {
+  RESTAURANT,
+  FOOD_SCOOTER
+}
