+++ src/app/store/promotion.ts
@@ -1,2 +1,9 @@
+import { PromotionType } from './promotion-type';
+
 export interface Promotion {
+  id: number,
+  startDate: Date,
+  endDate: Date,
+  type: PromotionType,
+  discount: number
 }
