+++ src/app/store/summaries/generalSummary.ts
