+++ src/app/store/summaries/locationSummary.ts
