+++ src/app/store/summaries/promotionSummary.ts
