+++ src/app/store/summaries/restaurantSummary.ts
